+++ benchmark/PackageConfigBench.php
@@ -9,15 +9,15 @@
 
 namespace InteropBench\Config;
 
-use InteropTest\Config\TestAsset\ConnectionMandatoryConfiguration;
+use InteropTest\Config\TestAsset\ConnectionConfiguration;
 
-class RequiresMandatoryOptionsBench extends BaseCase
+class PackageConfigBench extends BaseCase
 {
     /**
      * @inheritdoc \InteropBench\Config\BaseCase::getFactoryClass
      */
     protected function getFactoryClass()
     {
-        return new ConnectionMandatoryConfiguration();
+        return new ConnectionConfiguration();
     }
 }