+++ benchmark/RequiresMandatoryOptionsContainerIdBench.php
@@ -1,10 +1,10 @@
 <?php
 /**
- * Sake
+ * Sandro Keil (https://sandro-keil.de)
  *
  * @link      http://github.com/sandrokeil/interop-config for the canonical source repository
- * @copyright Copyright (c) 2015 Sandro Keil
- * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.txt New BSD License
+ * @copyright Copyright (c) 2015-2016 Sandro Keil
+ * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.md New BSD License
  */
 
 namespace InteropBench\Config;
@@ -13,6 +13,9 @@ use InteropTest\Config\TestAsset\ConnectionMandatoryContainerIdConfiguration;
 
 class RequiresMandatoryOptionsContainerIdBench extends BaseCase
 {
+    /**
+     * @inheritdoc \InteropBench\Config\BaseCase::getFactoryClass
+     */
     protected function getFactoryClass()
     {
         return new ConnectionMandatoryContainerIdConfiguration();