+++ src/ConfigurationTrait.php
@@ -1,118 +1,116 @@
 <?php
 /**
- * Sake
+ * Sandro Keil (https://sandro-keil.de)
  *
  * @link      http://github.com/sandrokeil/interop-config for the canonical source repository
- * @copyright Copyright (c) 2015 Sandro Keil
- * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.txt New BSD License
+ * @copyright Copyright (c) 2015-2016 Sandro Keil
+ * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.md New BSD License
  */
 
 namespace Interop\Config;
 
 use ArrayAccess;
+use Interop\Config\Exception;
 
 /**
- * ConfigurationTrait which retrieves options from configuration
+ * ConfigurationTrait which retrieves options from configuration, see interface \Interop\Config\RequiresConfig
  *
- * Use this trait if you want to retrieve options from a configuration and optional to perform a mandatory option check.
- * Default options are merged and overridden of the provided options.
+ * This trait is a implementation of \Interop\Config\RequiresConfig. Retrieves options from a configuration and optional
+ * to perform a mandatory option check. Default options are merged and overridden of the provided options.
  */
 trait ConfigurationTrait
 {
     /**
-     * @see \Interop\Config\RequiresConfig::vendorName
+     * @inheritdoc \Interop\Config\RequiresConfig::dimensions
      */
-    abstract public function vendorName();
+    abstract public function dimensions();
 
     /**
-     * @see \Interop\Config\RequiresConfig::packageName
+     * Checks if options are available depending on implemented interfaces and checks that the retrieved options are an
+     * array or have implemented \ArrayAccess.
+     *
+     * The RequiresConfigId interface is supported.
+     *
+     * @param array|ArrayAccess $config Configuration
+     * @param string|null $configId Config name, must be provided if factory uses RequiresConfigId interface
+     * @return bool True if options are available, otherwise false
      */
-    abstract public function packageName();
-
-    /**
-     * @see \Interop\Config\RequiresConfig::canRetrieveOptions
-     */
-    public function canRetrieveOptions($config)
+    public function canRetrieveOptions($config, $configId = null)
     {
-        if (!is_array($config) && !$config instanceof ArrayAccess) {
-            return false;
-        }
-        $id = null;
+        $dimensions = $this->dimensions();
 
-        if ($this instanceof RequiresContainerId) {
-            $id = $this->containerId();
+        if ($this instanceof RequiresConfigId) {
+            $dimensions[] = $configId;
         }
-        $vendorName = $this->vendorName();
-        $packageName = $this->packageName();
 
-        if (!isset($config[$vendorName][$packageName])
-            || (null !== $id && !isset($config[$vendorName][$packageName][$id]))
-        ) {
-            return false;
-        }
-        $options = $config[$vendorName][$packageName];
+        foreach ($dimensions as $dimension) {
+            if ((!is_array($config) && !$config instanceof ArrayAccess)
+                || (!isset($config[$dimension]) && $this instanceof RequiresMandatoryOptions)
+                || (!isset($config[$dimension]) && !$this instanceof ProvidesDefaultOptions)
+            ) {
+                return false;
+            }
+            if (!isset($config[$dimension]) && $this instanceof ProvidesDefaultOptions) {
+                return true;
+            }
 
-        if (null !== $id) {
-            $options = $options[$id];
+            $config = $config[$dimension];
         }
-        return is_array($options) || $options instanceof ArrayAccess;
+        return is_array($config) || $config instanceof ArrayAccess;
     }
 
     /**
-     * @see \Interop\Config\RequiresConfig::options
+     * Returns options based on dimensions() like [vendor][package] and can perform mandatory option checks if
+     * class implements RequiresMandatoryOptions. If the ProvidesDefaultOptions interface is implemented, these options
+     * must be overridden by the provided config. If you want to allow configurations for more then one instance use
+     * RequiresConfigId interface.
+     *
+     * The RequiresConfigId interface is supported.
+     *
+     * @param array|ArrayAccess $config Configuration
+     * @param string $configId Config name, must be provided if factory uses RequiresConfigId interface
+     * @return array|ArrayAccess
+     * @throws Exception\InvalidArgumentException If the $configId parameter is provided but factory does not support it
+     * @throws Exception\UnexpectedValueException If the $config parameter has the wrong type
+     * @throws Exception\OptionNotFoundException If no options are available
+     * @throws Exception\MandatoryOptionNotFoundException If a mandatory option is missing
      */
-    public function options($config)
+    public function options($config, $configId = null)
     {
-        if (!is_array($config) && !$config instanceof ArrayAccess) {
+        $options = $config;
+        $dimensions = $this->dimensions();
+
+        if ($this instanceof RequiresConfigId) {
+            $dimensions[] = $configId;
+        } elseif ($configId !== null) {
             throw new Exception\InvalidArgumentException(
-                sprintf('Provided parameter $config  must either be of type "array" or implement "\ArrayAccess".')
+                sprintf('The factory "%s" does not support multiple instances.', __CLASS__)
             );
         }
-        $id = null;
 
-        if ($this instanceof RequiresContainerId) {
-            $id = $this->containerId();
-        }
-        $vendorName = $this->vendorName();
-        $packageName = $this->packageName();
+        // get configuration for provided dimensions
+        foreach ($dimensions as $dimension) {
+            if (!is_array($options) && !$options instanceof ArrayAccess) {
+                throw Exception\UnexpectedValueException::invalidOptions($dimensions, $dimension);
+            }
 
-        // this is the fastest way to determine a configuration error (performance)
-        if (!isset($config[$vendorName][$packageName][$id])) {
-            if (!isset($config[$vendorName][$packageName])) {
-                if (empty($config[$vendorName])) {
-                    throw new Exception\OutOfBoundsException(
-                        sprintf('No vendor configuration "%s" available', $vendorName)
-                    );
+            if (!isset($options[$dimension])) {
+                if (!$this instanceof RequiresMandatoryOptions && $this instanceof ProvidesDefaultOptions) {
+                    break;
                 }
-                throw new Exception\OptionNotFoundException(
-                    sprintf('No options set for configuration "' . "['%s']['%s']", $vendorName, $packageName)
-                );
+                throw Exception\OptionNotFoundException::missingOptions($this, $dimension, $configId);
             }
-            if (null !== $id) {
-                throw new Exception\OptionNotFoundException(
-                    sprintf('No options set for configuration "' . "['%s']['%s']['%s']", $vendorName, $packageName, $id)
-                );
-            }
+            $options = $options[$dimension];
         }
-        $options = $config[$vendorName][$packageName];
 
-        if (null !== $id) {
-            $options = $options[$id];
-        }
         if (!is_array($options) && !$options instanceof ArrayAccess) {
-            throw new Exception\UnexpectedValueException(
-                sprintf(
-                    'Options of configuration ' . "['%s']['%s']%s"
-                    . ' must either be of type "array" or implement "\ArrayAccess".',
-                    $this->vendorName(),
-                    $this->packageName(),
-                    $id ? '["' . $id . '""]' : ''
-                )
-            );
+            throw Exception\UnexpectedValueException::invalidOptions($this->dimensions());
         }
+
         if ($this instanceof RequiresMandatoryOptions) {
             $this->checkMandatoryOptions($this->mandatoryOptions(), $options);
         }
+
         if ($this instanceof ProvidesDefaultOptions) {
             $options = array_replace_recursive($this->defaultOptions(), $options);
         }
@@ -124,14 +122,15 @@ trait ConfigurationTrait
      * an empty array will be returned.
      *
      * @param array|ArrayAccess $config Configuration
+     * @param string $configId Config name, must be provided if factory uses RequiresConfigId interface
      * @return array|ArrayAccess options Default options or an empty array
      */
-    public function optionsWithFallback($config)
+    public function optionsWithFallback($config, $configId = null)
     {
         $options = [];
 
-        if ($this->canRetrieveOptions($config)) {
-            $options = $this->options($config);
+        if ($this->canRetrieveOptions($config, $configId)) {
+            $options = $this->options($config, $configId);
         }
         if (empty($options) && $this instanceof ProvidesDefaultOptions) {
             $options = $this->defaultOptions();
@@ -155,22 +154,15 @@ trait ConfigurationTrait
                 $this->checkMandatoryOptions($mandatoryOption, $options[$key]);
                 return;
             }
+
             if (!$useRecursion && isset($options[$mandatoryOption])) {
                 continue;
             }
-            $id = null;
 
-            if ($this instanceof RequiresContainerId) {
-                $id = $this->containerId();
-            }
-
-            throw new Exception\MandatoryOptionNotFoundException(sprintf(
-                'Mandatory option "%s" was not set for configuration "' . "['%s']['%s']%s",
-                $useRecursion ? $key : $mandatoryOption,
-                $this->vendorName(),
-                $this->packageName(),
-                $id ? '["' . $id . '""]' : ''
-            ));
+            throw Exception\MandatoryOptionNotFoundException::missingOption(
+                $this->dimensions(),
+                $useRecursion ? $key : $mandatoryOption
+            );
         }
     }
 }