+++ src/RequiresConfig.php
@@ -1,18 +1,19 @@
 <?php
 /**
- * Sake
+ * Sandro Keil (https://sandro-keil.de)
  *
  * @link      http://github.com/sandrokeil/interop-config for the canonical source repository
- * @copyright Copyright (c) 2015 Sandro Keil
- * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.txt New BSD License
+ * @copyright Copyright (c) 2015-2016 Sandro Keil
+ * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.md New BSD License
  */
 
 namespace Interop\Config;
 
 use ArrayAccess;
+use Interop\Config\Exception;
 
 /**
- * RequiresConfig Interface
+ * RequiresConfig interface is the main interface to configure your instances via factories
  *
  * Use this interface if you want to retrieve options from a configuration and optional to perform a mandatory option
  * check. Default options are merged and overridden of the provided options.
@@ -20,38 +21,35 @@ use ArrayAccess;
 interface RequiresConfig
 {
     /**
-     * Returns the vendor name
+     * Returns the depth of the configuration array as a list. Can also be an empty array. For instance, the structure
+     * of the dimensions() method would be an array like
      *
-     * @return string
+     * <code>
+     *   return ['prooph', 'service_bus', 'command_bus'];
+     * </code>
+     *
+     * @return array|ArrayAccess
      */
-    public function vendorName();
+    public function dimensions();
 
     /**
-     * Returns the package name
-     *
-     * @return string
-     */
-    public function packageName();
-
-    /**
-     * Returns options based on [vendor][package][id] and can perform mandatory option checks if class implements
-     * MandatoryOptionsInterface. If the ProvidesDefaultOptions interface is implemented, these options must be
-     * overridden by the provided config. The RequiresContainerId interface is optional.
+     * Returns options based on dimensions() like [vendor][package] and can perform mandatory option checks if
+     * class implements RequiresMandatoryOptions. If the ProvidesDefaultOptions interface is implemented, these options
+     * must be overridden by the provided config. If you want to allow configurations for more then one instance use
+     * RequiresConfigId interface and add an optional second parameter named $configId. The ConfigurationTrait supports
+     * RequiresConfigId interface.
      *
      * <code>
      * return [
      *      // vendor name
-     *     'doctrine' => [
+     *     'prooph' => [
      *          // package name
-     *          'connection' => [
-     *             // container id, is optional
-     *             'orm_default' => [
-     *                 // mandatory options, is optional
-     *                 'driverClass' => 'Doctrine\DBAL\Driver\PDOMySql\Driver',
-     *                 'params' => [
-     *                     // default options, is optional
-     *                     'host' => 'localhost',
-     *                     'port' => '3306',
+     *          'service_bus' => [
+     *             // only one instance possible
+     *             'command_bus' => [
+     *                  // command bus factory options
+     *                 'router' => [
+     *                     'routes' => [],
      *                 ],
      *             ],
      *         ],
@@ -61,9 +59,8 @@ interface RequiresConfig
      *
      * @param array|ArrayAccess $config Configuration
      * @return array|ArrayAccess options
-     *
+     * @throws Exception\InvalidArgumentException If the $configId parameter is provided but factory does not support it
      * @throws Exception\UnexpectedValueException If the $config parameter has the wrong type
-     * @throws Exception\OutOfBoundsException If vendor name was not found
      * @throws Exception\OptionNotFoundException If no options are available
      * @throws Exception\MandatoryOptionNotFoundException If a mandatory option is missing
      */
@@ -71,7 +68,7 @@ interface RequiresConfig
 
     /**
      * Checks if options are available depending on implemented interfaces and checks that the retrieved options are an
-     * array or have implemented \ArrayAccess.
+     * array or have implemented \ArrayAccess. The ConfigurationTrait supports RequiresConfigId interface.
      *
      * @param array|ArrayAccess $config Configuration
      * @return bool True if options are available, otherwise false