+++ test/TestAsset/ConnectionConfiguration.php
@@ -1,10 +1,10 @@
 <?php
 /**
- * Sake
+ * Sandro Keil (https://sandro-keil.de)
  *
  * @link      http://github.com/sandrokeil/interop-config for the canonical source repository
- * @copyright Copyright (c) 2014 - 2015 Sandro Keil
- * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.txt New BSD License
+ * @copyright Copyright (c) 2015-2016 Sandro Keil
+ * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.md New BSD License
  */
 
 namespace InteropTest\Config\TestAsset;
@@ -16,13 +16,11 @@ class ConnectionConfiguration implements RequiresConfig
 {
     use ConfigurationTrait;
 
-    public function vendorName()
+    /**
+     * @interitdoc
+     */
+    public function dimensions()
     {
-        return 'doctrine';
-    }
-
-    public function packageName()
-    {
-        return 'connection';
+        return ['doctrine', 'connection'];
     }
 }