+++ test/TestAsset/PlainConfiguration.php
@@ -10,10 +10,9 @@
 namespace InteropTest\Config\TestAsset;
 
 use Interop\Config\ConfigurationTrait;
-use Interop\Config\RequiresConfigId;
-use Interop\Config\RequiresMandatoryOptions;
+use Interop\Config\RequiresConfig;
 
-class ConnectionMandatoryRecursiveContainerIdConfiguration implements RequiresConfigId, RequiresMandatoryOptions
+class PlainConfiguration implements RequiresConfig
 {
     use ConfigurationTrait;
 
@@ -22,14 +21,6 @@ class ConnectionMandatoryRecursiveContainerIdConfiguration implements RequiresCo
      */
     public function dimensions()
     {
-        return ['doctrine', 'connection'];
-    }
-
-    /**
-     * @interitdoc
-     */
-    public function mandatoryOptions()
-    {
-        return ['params' => ['user', 'dbname'], 'driverClass'];
+        return [];
     }
 }